refactor(channels): 提取公用常量和工具函数到 utils
