+++ src/renderer/pages/settings/channels/components/WeChatConfigForm.tsx
@@ -11,8 +11,7 @@ import GeminiModelSelector from '@/renderer/pages/settings/channels/components/G
 import { CHANNEL_DEFAULT_AGENT_BACKEND, type AcpBackendAll } from '@/types/acpTypes';
 import { useAppMode } from '@/renderer/hooks/useAppMode';
 import type { GeminiModelSelection } from '../hooks/useGeminiModelSelection';
-
-const WECHAT_GUIDE_URL = 'https://sudowork.sudoprivacy.com/guides/weixin-clawbot.html';
+import { WECHAT_GUIDE_URL } from '../utils';
 
 /**
  * Preference row component (matches DingTalk pattern)
